o  started updating Continuum facade and impl for IDs.
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497764 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/buildqueue/BuildProjectTask.java
@@ -28,9 +28,9 @@ import org.codehaus.plexus.taskqueue.Task;
 public class BuildProjectTask
     implements Task
 {
-    private int projectId;
+    private long projectId;
 
-    private int buildDefinitionId;
+    private long buildDefinitionId;
 
     private long timestamp;
 
@@ -38,7 +38,7 @@ public class BuildProjectTask
 
     private long maxExecutionTime;
 
-    public BuildProjectTask( int projectId, int buildDefinitionId, int trigger )
+    public BuildProjectTask( long projectId, long buildDefinitionId, int trigger )
     {
         this.projectId = projectId;
 
@@ -49,12 +49,12 @@ public class BuildProjectTask
         this.trigger = trigger;
     }
 
-    public int getProjectId()
+    public long getProjectId()
     {
         return projectId;
     }
 
-    public int getBuildDefinitionId()
+    public long getBuildDefinitionId()
     {
         return buildDefinitionId;
     }
